+++ main/pid.h
@@ -2,12 +2,12 @@
 struct PIDParams;
 
 typedef struct PIDParams {
-    double kp;
-    double ki;
-    double kd;
+    float kp;
+    float ki;
+    float kd;
 } PIDParams;
  
 // Function declarations
 // void PID_Init(PIDParams* params, double kp, double ki, double kd);
-double PID_Compute(PIDParams* params, double setpoint, double actual_value, double dt);
+float PID_Compute(PIDParams* params, float setpoint, float actual_value, float dt);
 void PID_Reset(PIDParams* params);
